Improve CFile data match and DrawError code match
- Make CManager::Init/Quit pure virtual (= 0) to generate correct
  __vt__8CManager layout with null function entries
- Change compactLayout type from int to u8 in DrawError, matching the
  target's clrlwi. byte-mask instruction

Results (main/file unit):
- 98/102 symbols at 100% match
- DrawError: 99.27% (remaining diff is register allocation)
- .rodata/.bss/.sdata2/extab/extabindex/.ctors: 100%
- .text: 99.73%, .data: 95.45%

The vtable RTTI pointer mismatch (target uses external RTTI refs
that no available compiler version reproduces) accounts for the
remaining data gap.

+++ src/file.cpp
@@ -185,7 +185,7 @@ retry:
 
         int hasScratchTexture = (int)Graphic.m_scratchTextureBuffer;
         hasScratchTexture = hasScratchTexture != 0;
-        int compactLayout = 0;
+        u8 compactLayout = 0;
         if (hasScratchTexture && usingFallbackFont == 0)
         {
             compactLayout = 1;
